added connection retries
