+++ tests/conftest.py
@@ -20,11 +20,18 @@ You should have received a copy of the GNU Lesser General Public License along w
 If not, see <https://www.gnu.org/licenses/>.
 """
 
+
 import os
+import time
 import pytest
+import multiprocessing
+import rpyc
 from PySide2 import QtWidgets
+from PySide2.QtCore import QTimer
 from qudi.core import application
 from qudi.util.yaml import yaml_load
+
+
 
 CONFIG = os.path.join(os.getcwd(),'tests/test.cfg')
 
@@ -91,3 +98,90 @@ def hardware_modules(config):
     base = 'hardware'
     modules = list(config[base].keys())
     return modules
+
+CONFIG = os.path.join(os.getcwd(),'tests/test.cfg')
+
+
+def run_qudi(timeout=150000):
+    """
+    Runs a Qudi instance with a timer.
+
+    Parameters
+    ----------
+    timeout : int, optional
+        timeout for the Qudi session in milliseconds, by default 150000.
+    """
+    app_cls = QtWidgets.QApplication
+    app = app_cls.instance()
+    if app is None:
+        app = app_cls()
+    qudi_instance = application.Qudi.instance()
+    if qudi_instance is None:
+        qudi_instance = application.Qudi(config_file=CONFIG)
+    QTimer.singleShot(timeout, qudi_instance.quit)
+    qudi_instance.run()
+
+
+@pytest.fixture(scope='module')
+def start_qudi_process():
+    """
+    Fixture that starts the Qudi process and ensures it's running before returning.
+    """
+    qudi_process = multiprocessing.Process(target=run_qudi)
+    qudi_process.start()
+    time.sleep(10)
+    yield
+    qudi_process.join(timeout=10)
+    if qudi_process.is_alive():
+        qudi_process.terminate()
+
+def connect_with_retries(host, port, config, retries=5, delay=2):
+    """
+    Attempt to connect multiple times to the RPyC server.
+
+    Parameters
+    ----------
+    host    : str 
+        The server's hostname or IP address.
+    port    : int 
+        The port number to connect to.
+    retries : int 
+        Number of retries before giving up.
+    delay   : int or float
+        Delay in seconds between attempts.
+
+    Returns
+    -------
+    conn (rpyc.Connection)
+        The RPyC connection if successful.
+
+    Raises:
+        Exception: If all attempts fail.
+    """
+    attempt = 0
+    while attempt < retries:
+        try:
+            print(f"Attempt {attempt + 1} of {retries}...")
+            conn = rpyc.connect(host, port, config=config)
+            print("Connection successful!")
+            return conn  # Return the connection object if successful
+        except Exception as e:
+            print(f"Connection attempt {attempt + 1} failed: {e}")
+            attempt += 1
+            if attempt < retries:
+                print(f"Retrying in {delay} seconds...")
+                time.sleep(delay)
+            else:
+                print("All connection attempts failed.")
+                raise  # Re-raise the last exception after final attempt
+
+@pytest.fixture(scope='module')
+def remote_instance(start_qudi_process):
+    """
+    Fixture that connects to the running Qudi ipython kernel through rpyc client and returns the client instance.
+    """
+    time.sleep(5)
+    conn = connect_with_retries(host="localhost", port=18861, config={'sync_request_timeout': 20})
+    root = conn.root
+    qudi_instance = root._qudi
+    return qudi_instance
